pyproject.toml
[tool.poetry]
name = "pytorch-ocr"
version = "0.1.0"
description = "Simple framework for OCR using PyTorch."
authors = ["Gabriel Dornelles", "Nick To <nick@nickto.net>"]
license = "MIT"
readme = "README.md"
packages = [{include = "pytorch_ocr"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
torch = "1.12.0"
torchvision = "0.13.0"
albumentations = "1.2.1"
numpy = "^1.25"
rich = "13.0.0"
hydra-core = "^1.3.2"
matplotlib = "3.5.3"
pytest = "^7.4.2"


[tool.black]
line-length = 119
target-version = ['py37', 'py38', 'py39', 'py310']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | venv
  | docs
  | _build
  | buck-out
  | build
  | dist
)/
'''
